% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Typicalsig.R
\name{calcPrevalence}
\alias{calcPrevalence}
\title{Calculate the prevalence of species and genera in any subset of cureatedMetagenomicData}
\usage{
calcPrevalence(df, rank, prevalencecutoff = 1e-06, thresholdcutoff = 0)
}
\arguments{
\item{df}{Any filtered subset of the rows of the sampleMetadata object from curatedMetagenomicData}

\item{rank}{taxonomic rank to calculate prevalences for. Must be one of "phylum", "class",
"order", "family", "genus", or "species"}

\item{prevalencecutoff}{Minimum prevalence of taxa to include in the output (default: 0.000001)}

\item{thresholdcutoff}{Minimum relative abundance to consider as present for the prevalence calculation.
This is a percentage out of 100. (default: 0)}
}
\value{
A data.frame containing names and prevalences at the taxonomic rank requested
}
\description{
Calculate the prevalence of species and genera in any subset of cureatedMetagenomicData
}
\examples{
library(dplyr)
healthy <- sampleMetadata \%>\%
   filter(disease == 'healthy') \%>\%
   select(where( ~ !all(is.na(.x)))) \%>\%
   head(100)
calcPrevalence(healthy, rank = "phylum")
}
